Refactored tap message classes.
These needed to be more readable and were horribly coded (by me). I
started seeing issues with conversion from bytes on the wire to
variables in spy so I changed how this was done.

Reviewed-on: http://review.couchbase.org/9712
Tested-by: Matt Ingenthron <matt@couchbase.com>

+++ src/main/java/net/spy/memcached/tapmessage/MessageBuilder.java
@@ -18,7 +18,7 @@ public class MessageBuilder {
 		message.setFlags(TapFlag.DUMP);
 	}
 
-	public void specifyVbuckets(int[] vbucketlist) {
+	public void specifyVbuckets(short[] vbucketlist) {
 		message.setVbucketlist(vbucketlist);
 		message.setFlags(TapFlag.LIST_VBUCKETS);
 	}
@@ -31,7 +31,7 @@ public class MessageBuilder {
 		message.setFlags(TapFlag.KEYS_ONLY);
 	}
 
-	public void takeoverVbuckets(int[] vbucketlist) {
+	public void takeoverVbuckets(short[] vbucketlist) {
 		message.setVbucketlist(vbucketlist);
 		message.setFlags(TapFlag.TAKEOVER_VBUCKETS);
 	}
